Stable Trading Strategy  (#75)
* add inGivenOut swap to Trade Script, remove merge-hop

* fix lint

* fix documentation

* add constant sum prod TS with math and cleaned logexpmath includes

* add stable ts lib, swapExactAmountIn to trading script and swap tests

* Apply suggestions from code review


* Update contracts/scripts/TradeScript.sol


* remove accum from trading strategy swaps

* update comment

* fix measureGas script

* reuse helper for last token calculated amount

* change balances from uint256 to uin128, fix the way to calculate invariant relative error on stable strategy

* fix stable strategy invariant validation

+++ contracts/vault/IVault.sol
@@ -41,7 +41,7 @@ interface IVault {
     function getPoolTokenBalances(bytes32 poolId, address[] calldata tokens)
         external
         view
-        returns (uint256[] memory);
+        returns (uint128[] memory);
 
     function getPoolTokens(bytes32 poolId)
         external
