+++ contracts/vault/PoolRegistry.sol
@@ -117,12 +117,12 @@ abstract contract PoolRegistry is IVault, VaultAccounting, BConst, Lock, Logs {
         external
         override
         view
-        returns (uint256[] memory)
+        returns (uint128[] memory)
     {
-        uint256[] memory balances = new uint256[](tokens.length);
+        uint128[] memory balances = new uint128[](tokens.length);
 
         for (uint256 i = 0; i < tokens.length; ++i) {
-            balances[i] = uint256(_poolTokenBalance[poolId][tokens[i]].total());
+            balances[i] = _poolTokenBalance[poolId][tokens[i]].total();
         }
 
         return balances;
